DROID-2274 views update
